Updated Tweet files so GET tweets works on Postman
In Tweet entity, changed author from Integer to User and changed inReplyTo and repostOf from int to Tweet.

In TweetMapper, changed entityToDto so it takes a Tweet instead of an Optional<Tweet>. This allows information to be displayed in Postman (it displays null if Optional<Tweet> is used). Conversion from Oprional<Tweet> to Tweet occurs in the implementaiton class.